+++ tests/test.rs
@@ -1,12 +1,20 @@
 #![warn(rust_2018_idioms, single_use_lifetimes)]
 
-use dependabot_config::{v2, Dependabot};
+use dependabot_config::{v1, v2, Dependabot};
 use serde::{Deserialize, Serialize};
 
 #[test]
 fn test() {
     let v2_test_data = include_str!("fixtures/v2.yml");
     assert_eq!(from_str::<serde_yaml::Value>(v2_test_data).len(), 22);
+    for case in from_str::<v2::Dependabot>(v2_test_data) {
+        case.to_string();
+    }
+    for case in from_str::<serde_yaml::Value>(v2_test_data) {
+        let s = serde_yaml::to_string(&case).unwrap();
+        s.parse::<v2::Dependabot>().unwrap().to_string();
+        serde_yaml::to_string(&serde_yaml::from_str::<v2::Dependabot>(&s).unwrap()).unwrap();
+    }
     for case in from_str::<Dependabot>(v2_test_data) {
         case.to_string();
     }
@@ -15,13 +23,24 @@ fn test() {
         s.parse::<Dependabot>().unwrap().to_string();
         serde_yaml::to_string(&serde_yaml::from_str::<Dependabot>(&s).unwrap()).unwrap();
     }
-    for case in from_str::<v2::Dependabot>(v2_test_data) {
+
+    let v1_test_data = include_str!("fixtures/v1.yml");
+    assert_eq!(from_str::<serde_yaml::Value>(v1_test_data).len(), 21);
+    for case in from_str::<v1::Dependabot>(v1_test_data) {
         case.to_string();
     }
-    for case in from_str::<serde_yaml::Value>(v2_test_data) {
+    for case in from_str::<serde_yaml::Value>(v1_test_data) {
         let s = serde_yaml::to_string(&case).unwrap();
-        s.parse::<v2::Dependabot>().unwrap().to_string();
-        serde_yaml::to_string(&serde_yaml::from_str::<v2::Dependabot>(&s).unwrap()).unwrap();
+        s.parse::<v1::Dependabot>().unwrap().to_string();
+        serde_yaml::to_string(&serde_yaml::from_str::<v1::Dependabot>(&s).unwrap()).unwrap();
+    }
+    for case in from_str::<Dependabot>(v1_test_data) {
+        case.to_string();
+    }
+    for case in from_str::<serde_yaml::Value>(v1_test_data) {
+        let s = serde_yaml::to_string(&case).unwrap();
+        s.parse::<Dependabot>().unwrap().to_string();
+        serde_yaml::to_string(&serde_yaml::from_str::<Dependabot>(&s).unwrap()).unwrap();
     }
 
     let v2_registries_test_data = include_str!("fixtures/v2_registries.yml");